+++ @es-labs/node/auth/knex.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const knex = require('../services/db/objection').get().knex()
+const knex = require('../services/db/knex').get()
 const { JWT_REFRESH_STORE_NAME } = global.CONFIG
 
 // id field must be unique index 
@@ -13,11 +13,8 @@ const setToken = async (id, refresh_token) => {
 const getToken = async (id) => {
   try {
     // TBD use DB - maybe better to use DB since it is already being read
-    // const User = require('../models/User') // TBD to change this...
     // const user = await User.query().where('id', '=', id) // TBD FIX THIS
-    // if (user && !user[0].revoked && req.body) {
-    //   refreshToken = user[0].refreshToken
-    // }
+    // if (user && !user[0].revoked && req.body) refreshToken = user[0].refreshToken
     const rv = await knex(JWT_REFRESH_STORE_NAME).where({ id: id }).first()
     return rv.refresh_token
   } catch (e) {