+++ example-app/deploy/db/knex/migrations/20180927133438_create_users.js
@@ -37,5 +37,5 @@ exports.up = async (knex) => {
 }
 
 exports.down = async (knex) => {
-  await knex.schema.dropTable('users')
+  await knex.schema.dropTableIfExists('users')
 }